Moved build summary to ts
