+++ frontend/public/locales/en.json
@@ -39,6 +39,7 @@
     "favourites": "Favourites",
     "productList": "Product List",
     "showLess": "Show less",
-    "showMore": "Show More"
+    "showMore": "Show More",
+    "versions": "Versions"
   }
 }
