--- frontend/src/modules/summary/productList/buildSummary/styles/BuildSummaryStyles.js
@@ -1,8 +1,0 @@
-const BuildSummaryStyles = () => ({
-  versionsSelector: {
-    paddingTop: '16px',
-    paddingBottom: '16px',
-  },
-});
-
-export default BuildSummaryStyles;
