+++ frontend/src/modules/summary/productList/buildSummary/styles/BuildSummaryStyles.ts
@@ -1,0 +1,10 @@
+import { makeStyles, createStyles } from '@material-ui/core';
+
+export const useBuildListStyles = makeStyles(() =>
+  createStyles({
+    versionsSelector: {
+      paddingTop: '16px',
+      paddingBottom: '16px',
+    },
+  })
+);
